Add video render

+++ app/api/routers/users.py
@@ -12,7 +12,7 @@ async def get_me(user: UserDep) -> UserOut:
     return user
 
 
-@router.post("/")
+@router.post("/create")
 async def create_user(db_session: SessionDep, user_create: UserIn) -> str:
     user = await users_crud.get_user_by_email(db_session, user_create.email)
     if user:
@@ -37,5 +37,7 @@ async def get_user_by_email(db_session: SessionDep, email: EmailStr) -> UserOut:
 @router.post("/delete/{email}")
 async def delete_user_by_email(db_session: SessionDep, email: str) -> UserOut:
     user_db = await users_crud.get_user_by_email(db_session, email)
+    if not user_db:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
     await users_crud.delete_user(db_session, email)
     return user_db
